chore: Replace all usage of origial useTokenApprovalActions
