+++ src/components/forms/lock_actions/LockForm/components/LockPreviewModal/components/LockActions.vue
@@ -3,7 +3,6 @@ import {
   TransactionReceipt,
   TransactionResponse,
 } from '@ethersproject/abstract-provider';
-import { parseUnits } from '@ethersproject/units';
 import { format } from 'date-fns';
 import { computed, onBeforeMount, reactive, ref, watch } from 'vue';
 import { useI18n } from 'vue-i18n';
@@ -14,11 +13,9 @@ import useConfig from '@/composables/useConfig';
 import useEthers from '@/composables/useEthers';
 import useNumbers, { FNumFormats } from '@/composables/useNumbers';
 import { dateTimeLabelFor } from '@/composables/useTime';
-import useTokenApprovalActions from '@/composables/approvals/useTokenApprovalActions';
 import useTransactions from '@/composables/useTransactions';
 import useNetwork from '@/composables/useNetwork';
 import { balancerContractsService } from '@/services/balancer/contracts/balancer-contracts.service';
-import { configService } from '@/services/config/config.service';
 import useWeb3 from '@/services/web3/useWeb3';
 import { TokenInfo } from '@/types/TokenList';
 import { TransactionActionInfo } from '@/types/transactions';
@@ -26,6 +23,7 @@ import useVotingGauges from '@/composables/useVotingGauges';
 import { VeBalLockInfo } from '@/services/balancer/contracts/contracts/veBAL';
 import { ApprovalAction } from '@/composables/approvals/types';
 import { captureException } from '@sentry/browser';
+import useTokenApprovalActions from '@/composables/approvals/useTokenApprovalActions';
 
 /**
  * TYPES
@@ -76,11 +74,6 @@ const { networkConfig } = useConfig();
 const { getProvider, explorerLinks } = useWeb3();
 const { addTransaction } = useTransactions();
 const { txListener, getTxConfirmedAt } = useEthers();
-const { getTokenApprovalActionsForSpender } = useTokenApprovalActions(
-  ref<string[]>([props.lockablePoolTokenInfo.address]),
-  ref<string[]>([props.lockAmount]),
-  ApprovalAction.Locking
-);
 const { fNum } = useNumbers();
 const { totalVotes, unallocatedVotes } = useVotingGauges();
 const { networkSlug } = useNetwork();
@@ -96,6 +89,18 @@ const lockActions = props.lockType.map((lockType, actionIndex) => ({
 }));
 
 const actions = ref<TransactionActionInfo[]>([...lockActions]);
+
+const amountsToApprove = computed(() => [
+  {
+    address: props.lockablePoolTokenInfo.address,
+    amount: props.lockAmount,
+  },
+]);
+const { getTokenApprovalActions } = useTokenApprovalActions({
+  amountsToApprove,
+  spender: networkConfig.addresses.veBAL,
+  actionType: ApprovalAction.Locking,
+});
 
 /**
  * COMPUTED
@@ -211,16 +216,7 @@ watch(lockActionStatesConfirmed, () => {
  * LIFECYCLE
  */
 onBeforeMount(async () => {
-  const approvalAmount = parseUnits(
-    props.lockAmount,
-    props.lockablePoolTokenInfo.decimals
-  ).toString();
-
-  const approvalActions = await getTokenApprovalActionsForSpender(
-    configService.network.addresses.veBAL,
-    approvalAmount
-  );
-
+  const approvalActions = await getTokenApprovalActions();
   actions.value.unshift(...approvalActions);
 });
 </script>
